+++ src/core/contextType.ts
@@ -114,6 +114,7 @@ import type { Collision } from "../ecs/systems/Collision";
 import type { SystemPhase } from "../ecs/systems/systems";
 import type { GameObjEventNames, GameObjEvents } from "../events/eventMap";
 import type { KEvent, KEventController, KEventHandler } from "../events/events";
+import type { AppScope, SceneScope } from "../events/scopes";
 import type { SceneDef } from "../game/scenes";
 import type { anchorPt } from "../gfx/anchor";
 import type { DrawBezierOpt } from "../gfx/draw/drawBezier";
@@ -160,29 +161,30 @@ import type {
 } from "../math/math";
 import type { NavMesh } from "../math/navigationmesh";
 import type { Vec2 } from "../math/Vec2";
-import type {
-    Anchor,
-    BlendMode,
-    Canvas,
-    Comp,
-    CompList,
-    Cursor,
-    EmptyComp,
-    GameObj,
-    GetOpt,
-    KAPLAYPlugin,
-    Key,
-    KGamepad,
-    KGamepadButton,
-    KGamepadStick,
-    LoadFontOpt,
-    Mask,
-    MouseButton,
-    MusicData,
-    QueryOpt,
-    RNGValue,
-    Shape,
-    Tag,
+import {
+    type Anchor,
+    type BlendMode,
+    type Canvas,
+    type Comp,
+    type CompList,
+    type Cursor,
+    type EmptyComp,
+    type GameObj,
+    type GetOpt,
+    type KAPLAYOpt,
+    type KAPLAYPlugin,
+    type Key,
+    type KGamepad,
+    type KGamepadButton,
+    type KGamepadStick,
+    type LoadFontOpt,
+    type Mask,
+    type MouseButton,
+    type MusicData,
+    type QueryOpt,
+    type RNGValue,
+    type Shape,
+    type Tag,
 } from "../types";
 import type { TupleWithoutFirst } from "../utils/types";
 import type { Engine } from "./engine";
@@ -5560,7 +5562,12 @@ export interface KAPLAYCtx {
      */
     StateMachine: typeof StateMachine;
     /**
-     * Define a scene.
+     * This object serves 2 purposes:
+     *
+     * * When called as a function, it defines a new scene with the
+     *   name and the initializer function.
+     * * It can also be used to register scene-local event handlers
+     *   that will be automatically cancelled when the scene is left.
      *
      * @param name - The scene name.
      * @param def - The scene definition.
@@ -5569,18 +5576,29 @@ export interface KAPLAYCtx {
      * ```js
      * // define a scene
      * scene("game", () => {
-     * // ...
+     *     // ...
      * });
-     *
      * // get options
      * scene("game", (opts) => {
      *     debug.log(opts.level);
      * });
      * ```
+     * @example
+     * ```js
+     * scene("pauseMenu", () => {
+     *     scene.onKeyPress("tab", () => {
+     *         debug.log("go to next menu item");
+     *     });
+     *     scene.onKeyPress("esc", () => {
+     *         // go back to game and cancel menu events
+     *         popScene();
+     *     });
+     * });
+     * ```
      *
      * @group Scenes
      */
-    scene(name: string, def: SceneDef): void;
+    scene: SceneScope;
     /**
      * Go to a scene, passing all rest args to scene callback.
      *
@@ -5602,50 +5620,58 @@ export interface KAPLAYCtx {
     go(name: string, ...args: any): void;
 
     /**
-     * Push the current active scene to a stack and enters in the new scene
+     * Push the current active scene to a stack and then goes to the new scene
      *
      * @param id - The scene name.
-     * @param args - The args passed to the scene defition.
+     * @param args - The args passed to the scene definition.
      *
      * @example
      * ```js
-     *  add([
-     *    text("this is the first scene", {size: 32 }),
-     *    pos(center()),
-     *  ]);
-     * scene("main", () => {
-     *  add([
-     *    sprite("bean"),
-     *    pos(center()),
-     *  ]);
+     * scene("mainScene", () => {
+     *     add([
+     *         text("this is the first scene", { size: 32 }),
+     *         pos(center()),
+     *     ]);
+     * });
+     * scene("otherScene", () => {
+     *     add([
+     *         sprite("bean"),
+     *         pos(center()),
+     *     ]);
      * });
      *
-     * pushScene("main")
+     * pushScene("mainScene")
      * ```
      *
      * @since v3001.1
      * @group Scenes
+     *
+     * @see {@link popScene}
      */
     pushScene(id: string, ...args: unknown[]): void;
 
     /**
      * Pops the scene from the stack and set as current active scene.
      *
+     * Only works if the current scene was entered using {@link pushScene}.
+     *
      * @example
      * ```js
-     *  add([
-     *    text("this is the first scene", {size: 32 }),
-     *    pos(center()),
-     *  ]);
-     * scene("main", () => {
-     *  add([
-     *    sprite("bean"),
-     *    pos(center()),
-     *  ]);
+     * scene("mainScene", () => {
+     *     add([
+     *         text("this is the first scene", { size: 32 }),
+     *         pos(center()),
+     *     ]);
+     * });
+     * scene("otherScene", () => {
+     *     add([
+     *         sprite("bean"),
+     *         pos(center()),
+     *     ]);
      * });
      *
-     * go("mainScene");
-     * popScene();  // when triggered the text should appear on the center screen //
+     * pushScene("mainScene");
+     * popScene(); // return to the current scene
      * ```
      *
      * @since v3001.1
@@ -6283,6 +6309,21 @@ export interface KAPLAYCtx {
      * @group Debug
      */
     debug: Debug;
+    /**
+     * The app scope for creating global events that won't be automatically cancelled
+     * when the scene is changed.
+     *
+     * @example
+     * ```js
+     * app.onKeyPress("f", () => {
+     *     debug.log("i run in every scene from now on");
+     * });
+     *
+     * // will NOT cancel the event
+     * go("someScene");
+     * ```
+     */
+    app: AppScope;
     /**
      * Import a plugin.
      *